Reorganización de la estructura de archivos
Usando un `package.json` por directorio, node los reconoce como módulos
y nos permite hacer imports absolutos.
