+++ src/assets/themes.js
@@ -1,7 +1,7 @@
 // Para importar todos los themes desde un sólo lugar, así:
 //
 //   import { Light, Dark, Coso } from './themes'
-import Light from './Light'
-import Dark from './Dark'
+import Light from 'assets/themes/Light'
+import Dark from 'assets/themes/Dark'
 
 export { Light, Dark }
